# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "API Minimarket",
                 "description":  "DocumentaciÃ³n de la API del sistema Minimarket",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
...
    "paths":  {
...
                  "/api/inventario":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "Inventario"
                                                               ],
                                                      "summary":  "Obtener movimientos de inventario",
                                                      "description":  "Devuelve una lista de todos los movimientos",
                                                      "operationId":  "listarMovimientos",
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "Lista obtenida exitosamente",
                                                                                    "content":  {
                                                                                                    "*/*":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/CollectionModelEntityModelInventario"
                                                                                                                           }
                                                                                                            }
                                                                                                }
                                                                                }
                                                                    },
                                                      "security":  [
                                                                       {
                                                                           "bearerAuth":  [

                                                                                          ]
                                                                       }
                                                                   ]
                                                  },
                                          "post":  {
                                                       "tags":  [
                                                                    "Inventario"
                                                                ],
                                                       "summary":  "Registrar movimiento",
                                                       "description":  "Registra una entrada de mercaderÃ­a o auditorÃ­a",
                                                       "operationId":  "registrarMovimiento",
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/Inventario"
                                                                                                                               }
                                                                                                                }
                                                                                       },
                                                                           "required":  true
                                                                       },
                                                       "responses":  {
                                                                         "201":  {
                                                                                     "description":  "Movimiento registrado exitosamente",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/EntityModelInventario"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "bearerAuth":  [

                                                                                           ]
                                                                        }
                                                                    ]
                                                   }
                                      },
...
                  "/api/inventario/producto/{productoId}":  {
                                                                "get":  {
                                                                            "tags":  [
                                                                                         "Inventario"
                                                                                     ],
                                                                            "summary":  "Obtener movimientos por producto",
                                                                            "description":  "Devuelve los movimientos de un producto especÃ­fico",
                                                                            "operationId":  "listarPorProducto",
                                                                            "parameters":  [
                                                                                               {
                                                                                                   "name":  "productoId",
                                                                                                   "in":  "path",
                                                                                                   "required":  true,
                                                                                                   "schema":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int64"
                                                                                                              }
                                                                                               }
                                                                                           ],
                                                                            "responses":  {
                                                                                              "200":  {
                                                                                                          "description":  "Movimientos obtenidos exitosamente",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "$ref":  "#/components/schemas/CollectionModelEntityModelInventario"
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      }
                                                                                          },
                                                                            "security":  [
                                                                                             {
                                                                                                 "bearerAuth":  [

                                                                                                                ]
                                                                                             }
                                                                                         ]
                                                                        }
                                                            },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "Link":  {
                                                    "type":  "object",
                                                    "properties":  {
                                                                       "href":  {
                                                                                    "type":  "string"
                                                                                },
                                                                       "hreflang":  {
                                                                                        "type":  "string"
                                                                                    },
                                                                       "title":  {
                                                                                     "type":  "string"
                                                                                 },
                                                                       "type":  {
                                                                                    "type":  "string"
                                                                                },
                                                                       "deprecation":  {
                                                                                           "type":  "string"
                                                                                       },
                                                                       "profile":  {
                                                                                       "type":  "string"
                                                                                   },
                                                                       "name":  {
                                                                                    "type":  "string"
                                                                                },
                                                                       "templated":  {
                                                                                         "type":  "boolean"
                                                                                     }
                                                                   }
                                                },
                                       "Categoria":  {
                                                         "type":  "object",
                                                         "properties":  {
                                                                            "id":  {
                                                                                       "type":  "integer",
                                                                                       "format":  "int64"
                                                                                   },
                                                                            "nombre":  {
                                                                                           "type":  "string"
                                                                                       }
                                                                        }
                                                     },
                                       "Producto":  {
                                                        "type":  "object",
                                                        "properties":  {
                                                                           "id":  {
                                                                                      "type":  "integer",
                                                                                      "format":  "int64"
                                                                                  },
                                                                           "nombre":  {
                                                                                          "type":  "string"
                                                                                      },
                                                                           "precio":  {
                                                                                          "type":  "number",
                                                                                          "format":  "double"
                                                                                      },
                                                                           "stock":  {
                                                                                         "type":  "integer",
                                                                                         "format":  "int32"
                                                                                     },
                                                                           "categoria":  {
                                                                                             "$ref":  "#/components/schemas/Categoria"
                                                                                         }
                                                                       }
                                                    },
...
                                       "Inventario":  {
                                                          "type":  "object",
                                                          "properties":  {
                                                                             "id":  {
                                                                                        "type":  "integer",
                                                                                        "format":  "int64"
                                                                                    },
                                                                             "producto":  {
                                                                                              "$ref":  "#/components/schemas/Producto"
                                                                                          },
                                                                             "cantidad":  {
                                                                                              "type":  "integer",
                                                                                              "format":  "int32"
                                                                                          },
                                                                             "tipoMovimiento":  {
                                                                                                    "type":  "string",
                                                                                                    "minLength":  1
                                                                                                },
                                                                             "fechaMovimiento":  {
                                                                                                     "type":  "string",
                                                                                                     "format":  "date-time"
                                                                                                 }
                                                                         },
                                                          "required":  [
                                                                           "cantidad",
                                                                           "fechaMovimiento",
                                                                           "producto",
                                                                           "tipoMovimiento"
                                                                       ]
                                                      },
                                       "EntityModelInventario":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "id":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               },
                                                                                        "producto":  {
                                                                                                         "$ref":  "#/components/schemas/Producto"
                                                                                                     },
                                                                                        "cantidad":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int32"
                                                                                                     },
                                                                                        "tipoMovimiento":  {
                                                                                                               "type":  "string",
                                                                                                               "minLength":  1
                                                                                                           },
                                                                                        "fechaMovimiento":  {
                                                                                                                "type":  "string",
                                                                                                                "format":  "date-time"
                                                                                                            },
                                                                                        "_links":  {
                                                                                                       "$ref":  "#/components/schemas/Links"
                                                                                                   }
                                                                                    },
                                                                     "required":  [
                                                                                      "cantidad",
                                                                                      "fechaMovimiento",
                                                                                      "producto",
                                                                                      "tipoMovimiento"
                                                                                  ]
                                                                 },
...
                                       "CollectionModelEntityModelInventario":  {
                                                                                    "type":  "object",
                                                                                    "properties":  {
                                                                                                       "_embedded":  {
                                                                                                                         "type":  "object",
                                                                                                                         "properties":  {
                                                                                                                                            "inventarioList":  {
                                                                                                                                                                   "type":  "array",
                                                                                                                                                                   "items":  {
                                                                                                                                                                                 "$ref":  "#/components/schemas/EntityModelInventario"
                                                                                                                                                                             }
                                                                                                                                                               }
                                                                                                                                        }
                                                                                                                     },
                                                                                                       "_links":  {
                                                                                                                      "$ref":  "#/components/schemas/Links"
                                                                                                                  }
                                                                                                   }
                                                                                },
...
                                       "Links":  {
                                                     "type":  "object",
                                                     "additionalProperties":  {
                                                                                  "$ref":  "#/components/schemas/Link"
                                                                              }
                                                 }
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "AutenticaciÃ³n JWT usando Bearer Token",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}